requirements.txt
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.10
pypdf==4.0.1
faiss-cpu==1.7.4
python-dotenv==1.0.0
tiktoken==0.5.2
streamlit==1.29.0
numpy==1.24.3
pandas==2.0.3